Code Corrections
Velocity Chart :- It is useful during sprint planning meetings. This chart shows the amount of value delivered in each sprint, enabling you to predict the amount of work the team or an individual can get done in future sprints, to help us decide how much work we can feasibly commit to.

Estimation statistic :- The y-axis displays the statistic used for estimating stories, here we have used story points. Estimates can also be based on business value, hours, issue count, or any numeric field of your choice.
Assigned: The gray bar for each sprint shows the total estimate of all issues in the sprint when it begins.
Completed: The green bar in each sprint shows the total completed estimates when the sprint ends. 
Sprints: The x-axis displays the last sprints completed by the team. This data is used to calculate velocity.